Rename ViewModels to InputModels for Campaign and AdGroup
