+++ Web/Controllers/AdGroupController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Maximiz.InputModels.AdGroup;
 using Maximiz.Model.Entity;
 using Microsoft.AspNetCore.Mvc;
 
@@ -16,7 +17,7 @@ namespace Maximiz.Controllers
         }
 
         [HttpPost]
-        public IActionResult Create(AdGroup model)
+        public IActionResult Create(CreateAdGroupModel model)
         {
             // Test
             var x = model.Name;
