+++ Web/Views/AdGroup/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model  Maximiz.Model.Entity.AdGroup
+﻿@model  Maximiz.InputModels.AdGroup.CreateAdGroupModel
 
 @{
     ViewData["Title"] = "New Ad Group";
